new Games class & changes in GamePreview class
GamePreview class:
- add method searchGame(String)
- add method toString(List<GamePreview>)

Game class:
- moved searchGame(String) to GamePreview class

Games class:
- override toString() method
- moved saveToFile() method from main to this class
- moved readFromFile() method from main to this class
